% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EastingNorthing.R
\name{EastingNorthing}
\alias{EastingNorthing}
\title{Get Easting and Northing coordinates from DEFRA}
\usage{
EastingNorthing(IDs)
}
\arguments{
\item{IDs}{An alphanumeric string (or vector of strings) containing the UK AIR ID defined by DEFRA.}
}
\value{
A named vector containing Easting and Northing coordinates.
}
\description{
This function takes as input the UK AIR ID and returns Easting and Northing coordinates (British National Grid, EPSG:27700).
}
\examples{
# EastingNorthing("UKA12536")
# EastingNorthing(c("UKA15910", "UKA15956", "UKA16663", "UKA16097"))

}

