% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteID.R
\name{getSiteID}
\alias{getSiteID}
\title{Get site IDs for DEFRA stations}
\usage{
getSiteID(IDs)
}
\arguments{
\item{IDs}{An alphanumeric string (or vector of strings) containing the UK AIR ID defined by DEFRA.}
}
\value{
A named vector containing the site IDs.
}
\description{
This function fetches the catalogue of monitoring stations from DEFRA's website.
}
\examples{
# getSiteID("UKA00399")

}

