% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdefra-package.R
\docType{package}
\name{rdefra-package}
\alias{rdefra}
\alias{rdefra-package}
\title{rdefra: Interact with the UK AIR Pollution Database from DEFRA}
\description{
This packages allows to retrieve air pollution data from the Air Information Resource (UK-AIR, \url{https://uk-air.defra.gov.uk/}) of the Department for Environment, Food and Rural Affairs (DEFRA) in the United Kingdom. UK-AIR does not provide a public API for programmatic access to data, therefore this package scrapes the HTML pages to get relevant information. The package is described in Vitolo et al. (2016) "rdefra: Interact with the UK AIR Pollution Database from DEFRA" \doi{10.21105/joss.00051}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/rdefra/}
  \item \url{https://github.com/ropensci/rdefra}
  \item Report bugs at \url{https://github.com/ropensci/rdefra/issues}
}

}
\author{
\strong{Maintainer}: Claudia Vitolo \email{cvitolodev@gmail.com} (\href{https://orcid.org/0000-0002-4252-1176}{ORCID}) (Claudia is the original creator of the package)

Authors:
\itemize{
  \item Cheryl Lim \email{cheryl.academic@gmail.com}
}

Other contributors:
\itemize{
  \item Andrew Russell (This package is part of the KEHRA project and Andrew supervised the technical/software development.) [contributor]
  \item Allan Tucker (This package is part of the KEHRA project and Allan supervised the technical/software development.) [contributor]
  \item Maëlle Salmon (Maëlle Salmon reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/68) [contributor]
  \item Hao Zhu (Hao Zhu reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/68) [contributor]
}

}
\keyword{internal}
