% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdefra-package.R
\docType{package}
\name{rdefra}
\alias{rdefra}
\alias{rdefra-package}
\title{rdefra: Interact with the UK AIR Pollution Database from DEFRA}
\description{
Get data from DEFRA's UK-AIR website (see \url{https://uk-air.defra.gov.uk/}). It basically scraps the HTML content.
}

