% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukair_catalogue.R
\name{ukair_catalogue}
\alias{ukair_catalogue}
\title{Get DEFRA UK-AIR stations metadata}
\usage{
ukair_catalogue(site_name = "", pollutant = 9999, group_id = 9999,
  closed = "true", country_id = 9999, region_id = 9999,
  location_type = 9999)
}
\arguments{
\item{site_name}{This is the name of a specific site (by default this is left blank to get info on all the available sites.}

\item{pollutant}{This is a number from 1 to 10. Default is 9999, which means all the pollutants.}

\item{group_id}{This is the identification number of a group of stations. Default is 9999 which means all available networks.}

\item{closed}{This is "true" to include closed stations, "false" otherwise.}

\item{country_id}{This is the identification number of the country, it can be a number from 1 to 6. Default is 9999, which means all the countries.}

\item{region_id}{This is the identification number of the region. 1 = Aberdeen City, etc. (for the full list see \url{https://uk-air.defra.gov.uk/}). Default is 9999, which means all the local authorities.}

\item{location_type}{This is the identification number of the location. Default is 9999, which means all the location types.}
}
\value{
A named vector containing Easting and Northing coordinates.
}
\description{
This function fetches the catalogue of monitoring stations from DEFRA's website.
}
\details{
\code{Pollutant} is defined based on the following convention: 1 = Ozone (O3), 2 = Nitrogen oxides (NOx), 3 = Carbon monoxide (CO), 4 = Sulphur dioxide (SO2), 5 = Particulate Matter (PM10), 6 = Particulate Matter (PM2.5), 7 = PAHs, 8 = Metals in PM10, 9 = Benzene, 10 = Black Carbon.
\code{group_id} is defined based on the following convention: 1 = UKEAP: Precip-Net, 2 = Air Quality Strategy Pollutants, 3 = Ammonia and Nitric Acid, 4 = Automatic Urban and Rural Monitoring Network (AURN), 5 = Dioxins and Furans, 6 = Black Smoke & SO2, 7 = Automatic Hydrocarbon Network, 8 = Heavy Metals, 9 = Nitrogen Dioxide Diffusion Tube, 10 = PAH Andersen, 11 = Particle Size Composition, 12 = PCBs, 13 = TOMPs, 14 = Non-Automatic Hydrocarbon Network, 15 = 1,3-Butadiene Diffusion Tube, 16 = Black Carbon, 17 = Automatic Urban and Rural Monitoring Network (AURN), 18 = Defra NO2 Diffusion Tube, 19 = PAH Digitel (solid phase), 20 = PAH Digitel (solid+vapour), 21 = PAH Deposition, 22 = Particle size and number, 23 = Rural Automatic Mercury network, 24 = Urban Sulphate, 25 = UKEAP: Rural NO2, 26 = Automatic Urban and Rural Monitoring Network (AURN), 27 = UKEAP: National Ammonia Monitoring Network, 28 = UKEAP: Acid Gases & Aerosol Network, 29 = Particle Speciation (MARGA), 30 = UKEAP: Historic Aerosol measurements.
\code{country_id} is defined based on the following convention: 1 = England, 2 = Wales, 3 = Scotland, 4 = Northern Ireland, 5 = Republic of Ireland, 6 = Channel Islands.
}
\examples{
 \dontrun{
 ukair_catalogue()
 }

}

