% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukair_get_hourly_data.R
\name{ukair_get_hourly_data}
\alias{ukair_get_hourly_data}
\title{Get hourly data for DEFRA stations}
\usage{
ukair_get_hourly_data(site_id = NULL, years = NULL, keepUnits = FALSE)
}
\arguments{
\item{site_id}{This is the ID of a specific site.}

\item{years}{Years for which data should be downloaded.}

\item{keepUnits}{logical that if set to TRUE returns a column with unit after each column with measurements (this parameter is set to FALSE by default).}
}
\value{
A data.frame containing hourly pollution data.
}
\description{
This function fetches hourly data from DEFRA's air pollution monitoring stations.
}
\details{
The measurements are generally in ugm-3 (micrograms per cubic metre). Please double check the units before using these data.
}
\examples{
 \dontrun{
 ukair_get_hourly_data("ABD", 2014)
 ukair_get_hourly_data("ABD", 2014:2016)
 }

}

