% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukair_get_site_id.R
\name{ukair_get_site_id}
\alias{ukair_get_site_id}
\title{Get site identification numbers for DEFRA stations}
\usage{
ukair_get_site_id(id_s)
}
\arguments{
\item{id_s}{An alphanumeric string (or vector of strings) containing the UK
AIR ID defined by DEFRA.}
}
\value{
A named vector containing the site id_s.
}
\description{
Given the UK AIR ID (from the \code{ukair_catalogue()}), this
function fetches the catalogue of monitoring stations from DEFRA's website.
}
\examples{
 \dontrun{
 ukair_get_site_id("UKA00399")
 }

}
