% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukair_get_coordinates.R
\name{ukair_get_coordinates}
\alias{ukair_get_coordinates}
\title{Get Easting and Northing coordinates from DEFRA}
\usage{
ukair_get_coordinates(IDs)
}
\arguments{
\item{IDs}{An alphanumeric string (or vector of strings) containing the UK AIR ID defined by DEFRA.}
}
\value{
A named vector containing Easting and Northing coordinates.
}
\description{
This function takes as input the UK AIR ID and returns Easting and Northing coordinates (British National Grid, EPSG:27700).
}
\examples{
 \dontrun{
 ukair_get_coordinates("UKA12536")
 ukair_get_coordinates(c("UKA15910", "UKA15956", "UKA16663", "UKA16097"))
 }

}

